[ISSUE #9664]♻️Extract Tokio client request, one-way, and GO_AWAY orchestration
